+++ src/common/Token.php
@@ -11,8 +11,6 @@ class Token extends BaseObject implements TreeNodeInterface
     private $type;
     /** @var string Matched content of the token */
     private $content;
-    /** @var bool Whether then token is hidden with respect to `Symbol` definition */
-    private $isHidden = false;
     /** @var bool Whether then token is inline with respect to `Grammar` inline tokens */
     private $isInline = false;
     /** @var array|null Match data for the token given from `preg_match()` */
@@ -27,8 +25,6 @@ class Token extends BaseObject implements TreeNodeInterface
      * @param string $content Matched content of the token
      * @param array|null $match Match data for the token given from `preg_match()`
      * @param int|null $offset Position of the token in the input text
-     * @param bool $isHidden [since 1.4.0] Whether then token is hidden with respect to `Symbol`
-     * definition
      * @param bool $isInline [since 1.5.0] Whether then token is inline with respect to `Grammar`
      * inline tokens
      */
@@ -37,14 +33,12 @@ class Token extends BaseObject implements TreeNodeInterface
         string $content,
         ?array $match = null,
         ?int $offset = null,
-        bool $isHidden = false,
         bool $isInline = false
     ) {
         $this->type = $type;
         $this->content = $content;
         $this->match = $match;
         $this->offset = $offset;
-        $this->isHidden = $isHidden;
         $this->isInline = $isInline;
     }
 
@@ -64,16 +58,6 @@ class Token extends BaseObject implements TreeNodeInterface
     public function getContent(): string
     {
         return $this->content;
-    }
-
-    /**
-     * Whether then token is hidden with respect to `Symbol` definition
-     * @return bool
-     * @since 1.4.0
-     */
-    public function isHidden(): bool
-    {
-        return $this->isHidden;
     }
 
     /**